[HttpFoundation] Extend ClockMock to session storage tests
