+++ src/Symfony/Component/HttpFoundation/Tests/ResponseHeaderBagTest.php
@@ -14,20 +14,8 @@ namespace Symfony\Component\HttpFoundation\Tests;
 use Symfony\Component\HttpFoundation\ResponseHeaderBag;
 use Symfony\Component\HttpFoundation\Cookie;
 
-require_once __DIR__.'/ClockMock.php';
-
-class ResponseHeaderBagTest extends \PHPUnit_Framework_TestCase
+class ResponseHeaderBagTest extends ClockMockTestCase
 {
-    protected function setUp()
-    {
-        with_clock_mock(true);
-    }
-
-    protected function tearDown()
-    {
-        with_clock_mock(false);
-    }
-
     /**
      * @covers Symfony\Component\HttpFoundation\ResponseHeaderBag::allPreserveCase
      * @dataProvider provideAllPreserveCase
