+++ src/Symfony/Component/HttpFoundation/Tests/Session/Storage/Handler/MongoDbSessionHandlerTest.php
@@ -12,12 +12,13 @@
 namespace Symfony\Component\HttpFoundation\Tests\Session\Storage\Handler;
 
 use Symfony\Component\HttpFoundation\Session\Storage\Handler\MongoDbSessionHandler;
+use Symfony\Component\HttpFoundation\Tests\ClockMockTestCase;
 
 /**
  * @author Markus Bachmann <markus.bachmann@bachi.biz>
  * @requires extension mongo
  */
-class MongoDbSessionHandlerTest extends \PHPUnit_Framework_TestCase
+class MongoDbSessionHandlerTest extends ClockMockTestCase
 {
     /**
      * @var \PHPUnit_Framework_MockObject_MockObject
@@ -28,6 +29,8 @@ class MongoDbSessionHandlerTest extends \PHPUnit_Framework_TestCase
 
     protected function setUp()
     {
+        parent::setUp();
+
         $mongoClass = version_compare(phpversion('mongo'), '1.3.0', '<') ? 'Mongo' : 'MongoClient';
 
         $this->mongo = $this->getMockBuilder($mongoClass)
