+++ src/Symfony/Component/HttpFoundation/Tests/Session/Storage/MetadataBagTest.php
@@ -12,11 +12,12 @@
 namespace Symfony\Component\HttpFoundation\Tests\Session\Storage;
 
 use Symfony\Component\HttpFoundation\Session\Storage\MetadataBag;
+use Symfony\Component\HttpFoundation\Tests\ClockMockTestCase;
 
 /**
  * Test class for MetadataBag.
  */
-class MetadataBagTest extends \PHPUnit_Framework_TestCase
+class MetadataBagTest extends ClockMockTestCase
 {
     /**
      * @var MetadataBag
@@ -30,6 +31,7 @@ class MetadataBagTest extends \PHPUnit_Framework_TestCase
 
     protected function setUp()
     {
+        parent::setUp();
         $this->bag = new MetadataBag();
         $this->array = array(MetadataBag::CREATED => 1234567, MetadataBag::UPDATED => 12345678, MetadataBag::LIFETIME => 0);
         $this->bag->initialize($this->array);
@@ -39,6 +41,7 @@ class MetadataBagTest extends \PHPUnit_Framework_TestCase
     {
         $this->array = array();
         $this->bag = null;
+        parent::tearDown();
     }
 
     public function testInitialize()
